+++ mail/mail-common/src/errors/proton_mail_error.rs
@@ -1,6 +1,6 @@
 use super::mail_error_reason::*;
 use crate::actions::ActionError;
-use crate::draft::PackageError;
+use crate::draft::{Error, PackageError};
 use crate::errors::api_service_error::UserApiServiceError;
 use crate::errors::unexpected::Unexpected;
 use crate::{draft::Error as DraftError, AppError, MailContextError, MailboxError, SidebarError};
@@ -231,6 +231,10 @@ impl From<DraftError> for ProtonMailError {
             DraftError::SendMessage(package_error) => Self::from(package_error),
             DraftError::NoRecipients => {
                 Self::Reason(MailErrorReason::DraftReason(DraftErrorReason::NoRecipients))
+            }
+            DraftError::DeleteFailed => Self::Unexpected(Unexpected::Api),
+            Error::AlreadySent => {
+                Self::Reason(MailErrorReason::DraftReason(DraftErrorReason::AlreadySent))
             }
         }
     }