+++ mail/mail-common/tests/draft_send.rs
@@ -30,6 +30,7 @@ use proton_mail_common::{draft, MailContextError, MailUserContext};
 use proton_mail_ids::LocalMessageId;
 use proton_mail_test_utils::init::Params as TestParams;
 use proton_mail_test_utils::message_body::*;
+use proton_mail_test_utils::messages::TestDraftSendRequest;
 use proton_mail_test_utils::test_context::MailTestContext;
 use stash::orm::Model;
 use std::sync::Arc;
@@ -104,8 +105,9 @@ async fn basic_send_check() {
         DraftAttachmentKeyPackets::new(),
     )
     .await;
-    ctx.mock_send_draft_basic(
+    ctx.mock_send_draft(
         message.metadata.id.clone(),
+        default_mock_send_params(),
         sent_message.clone(),
         sent_conversation,
     )
@@ -326,6 +328,57 @@ async fn draft_save_failure_creates_send_result_with_correct_origin_when_used_be
     assert_eq!(send_result.origin, DraftSendResultOrigin::SaveBeforeSend);
 }
 
+#[tokio::test]
+async fn save_after_send_is_an_error() {
+    // Re-saving a draft after a queued send action is not allowed.
+    let ctx = MailTestContext::with_user_secret_and_user_id(
+        message_body_test_user_secret(),
+        UserId::from(TEST_USER_ID),
+    )
+    .await;
+    let params = draft_test_params();
+    let user_ctx = ctx.mail_user_context().await;
+
+    let mut message = message_body_test_message_simple();
+    message.metadata.to_list.push(MessageRecipient {
+        address: "foo@bar.com".to_string(),
+        is_proton: false,
+        name: "".to_string(),
+        group: None,
+    });
+
+    ctx.setup_user(params.clone()).await;
+    ctx.catch_all().await;
+    ctx.init_user(user_ctx.clone()).await;
+
+    // Create draft.
+    let mut draft = Draft::empty(user_ctx.user_stash()).await.unwrap();
+    draft
+        .to_list
+        .add_single(RecipientEntry {
+            email: "foo@bar.com".into(),
+            display_name: MaybeEmptyString(None),
+        })
+        .unwrap();
+    user_ctx
+        .with_queue(|queue| draft.save(queue))
+        .await
+        .unwrap();
+
+    // Save at least once so we can retrieve the message id.
+    let send_action = draft.to_send_action().unwrap();
+    user_ctx
+        .with_queue(|queue| send_action.queue(queue))
+        .await
+        .unwrap();
+
+    let result = user_ctx.with_queue(|queue| draft.save(queue)).await;
+    assert!(matches!(
+        result,
+        Err(MailContextError::Draft(draft::Error::AlreadySent))
+    ));
+}
+
 async fn send_fails_if_recipient_is_not_valid_impl(
     api_error_code: u32,
 ) -> (Arc<anyhow::Error>, LocalMessageId, Arc<MailUserContext>) {
@@ -421,6 +474,7 @@ fn draft_test_params_impl(mime_type: Option<MimeType>) -> TestParams {
     if let Some(mime_type) = mime_type {
         mail_settings.draft_mime_type = mime_type.into();
     }
+    mail_settings.delay_send_seconds = SEND_DELAY_SECONDS;
     let mut params = TestParams {
         user_info: Some(message_body_test_user_info()),
         addresses: message_body_test_addresses(),
@@ -471,6 +525,8 @@ fn draft_test_params_impl(mime_type: Option<MimeType>) -> TestParams {
     params
 }
 
+const SEND_DELAY_SECONDS: u32 = 60;
+
 fn expected_create_draft_params() -> DraftParams {
     let address = message_body_test_addresses();
     DraftParams {
@@ -491,5 +547,15 @@ fn expected_create_draft_params() -> DraftParams {
         draft_flags: 0,
         body: EncryptedDraft(String::new()),
         mime_type: MailSettings::default().draft_mime_type.into(),
+    }
+}
+
+fn default_mock_send_params() -> TestDraftSendRequest {
+    TestDraftSendRequest {
+        expiration_time: None,
+        expires_in: None,
+        auto_save_contacts: Some(true),
+        delay_seconds: Some(SEND_DELAY_SECONDS.into()),
+        delivery_time: None,
     }
 }