+++ mail/mail-uniffi/CHANGELOG.md
@@ -7,9 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased] - 2025-01-XX
 
+### Added
+
+  - [ET-1417] `Draft::discard` - Discards a draft from the composer.
+  - More logging for the html transformations.
+
 ### Fixed
 
- - Drafts are moved to outbox before being sent
+  - Drafts are moved to outbox before being sent.
+  - [ET-503] Drafts can not be updated after being sent.
 
 ## [0.33.0] - 2025-01-17